+++ crates/rpc/src/server/api.rs
@@ -1,3 +1,5 @@
+use std::net::SocketAddr;
+
 use miden_node_proto::{
     generated::{
         block_producer::api_client as block_producer_client,
@@ -31,37 +33,38 @@ use tonic::{
 };
 use tracing::{debug, info, instrument};
 
-use crate::{COMPONENT, config::RpcConfig};
+use crate::COMPONENT;
 
-// RPC API
+// RPC SERVICE
 // ================================================================================================
 
 type StoreClient = store_client::ApiClient<InterceptedService<Channel, OtelInterceptor>>;
 type BlockProducerClient =
     block_producer_client::ApiClient<InterceptedService<Channel, OtelInterceptor>>;
 
-pub struct RpcApi {
+pub struct RpcService {
     store: StoreClient,
     block_producer: BlockProducerClient,
 }
 
-impl RpcApi {
-    pub(super) async fn from_config(config: &RpcConfig) -> Result<Self, Error> {
-        let channel = tonic::transport::Endpoint::try_from(config.store_url.to_string())?
-            .connect()
-            .await?;
+impl RpcService {
+    pub(super) async fn new(
+        store_address: SocketAddr,
+        block_producer_address: SocketAddr,
+    ) -> Result<Self, Error> {
+        let store_url = format!("http://{store_address}");
+        let channel = tonic::transport::Endpoint::try_from(store_url)?.connect().await?;
         let store = store_client::ApiClient::with_interceptor(channel, OtelInterceptor);
-        info!(target: COMPONENT, store_endpoint = config.store_url.as_str(), "Store client initialized");
+        info!(target: COMPONENT, store_endpoint = %store_address, "Store client initialized");
 
-        let channel = tonic::transport::Endpoint::try_from(config.block_producer_url.to_string())?
-            .connect()
-            .await?;
+        let block_producer_url = format!("http://{block_producer_address}");
+        let channel = tonic::transport::Endpoint::try_from(block_producer_url)?.connect().await?;
         let block_producer =
             block_producer_client::ApiClient::with_interceptor(channel, OtelInterceptor);
 
         info!(
             target: COMPONENT,
-            block_producer_endpoint = config.block_producer_url.as_str(),
+            block_producer_endpoint = %block_producer_address,
             "Block producer client initialized",
         );
 
@@ -70,7 +73,7 @@ impl RpcApi {
 }
 
 #[tonic::async_trait]
-impl api_server::Api for RpcApi {
+impl api_server::Api for RpcService {
     #[instrument(
         target = COMPONENT,
         name = "rpc:check_nullifiers",